+++ mountpoint-s3/tests/cli.rs
@@ -126,18 +126,7 @@ fn bucket_name_and_directory_swapped() -> Result<(), Box<dyn std::error::Error>>
     let mut cmd = Command::cargo_bin("mount-s3")?;
 
     cmd.arg("test/dir").arg("my-bucket-name");
-    let error_message = "bucket argument should be a valid bucket name(only letters, numbers, . and -) or a valid ARN";
-    cmd.assert().failure().stderr(predicate::str::contains(error_message));
-
-    Ok(())
-}
-
-#[test]
-fn s3_uri_as_bucket_name() -> Result<(), Box<dyn std::error::Error>> {
-    let mut cmd = Command::cargo_bin("mount-s3")?;
-
-    cmd.arg("s3://test-bucket/").arg("test/dir");
-    let error_message = "bucket name should not be an s3:// URI";
+    let error_message = "the bucket must have a valid name (only letters, numbers, . and -) or a valid ARN";
     cmd.assert().failure().stderr(predicate::str::contains(error_message));
 
     Ok(())
@@ -336,6 +325,22 @@ fn fstab_cannot_use_read_only() -> Result<(), Box<dyn std::error::Error>> {
 
     cmd.arg("test-bucket").arg(dir.path()).arg("-o").arg("read-only");
     let error_message = "Cannot use 'read-only' flag when using fstab style arguments";
+    cmd.assert().failure().stderr(predicate::str::contains(error_message));
+
+    Ok(())
+}
+
+#[cfg(feature = "fstab")]
+#[test]
+fn fstab_cannot_use_multiple_prefixes() -> Result<(), Box<dyn std::error::Error>> {
+    let dir = assert_fs::TempDir::new()?;
+    let mut cmd = Command::cargo_bin("mount-s3")?;
+
+    cmd.arg("s3://test-bucket/prefix/")
+        .arg(dir.path())
+        .arg("-o")
+        .arg("prefix=foo/");
+    let error_message = "explicit prefix option not allowed with S3 URI";
     cmd.assert().failure().stderr(predicate::str::contains(error_message));
 
     Ok(())